requirements.txt
pandas>=1.5
numpy>=1.23
scipy>=1.10
statsmodels>=0.14
yfinance>=0.2
streamlit>=1.25
matplotlib>=3.7
seaborn>=0.12
scikit-learn>=1.2
tqdm>=4.65
pdfplumber>=0.10
fpdf2>=2.7
pyyaml>=6.0
python-dateutil>=2.8
plotly>=5.15
kaleido>=0.2
duckdb>=0.9
requests>=2.31
opencv-python-headless>=4.8
prefect>=2.14
websocket-client>=1.6
sentinelsat>=1.2
